review fixes 8
