+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/extension/ToBooleanConversionExt.scala
@@ -1,8 +1,8 @@
 package pl.touk.nussknacker.engine.extension
 
-import pl.touk.nussknacker.engine.api.generics.MethodTypeInfo
+import pl.touk.nussknacker.engine.api.typed.typing
 import pl.touk.nussknacker.engine.api.typed.typing.Typed
-import pl.touk.nussknacker.engine.definition.clazz.{ClassDefinitionSet, MethodDefinition, StaticMethodDefinition}
+import pl.touk.nussknacker.engine.definition.clazz.{ClassDefinitionSet, StaticMethodDefinition}
 
 import java.lang.{Boolean => JBoolean}
 
@@ -12,22 +12,17 @@ class ToBooleanConversionExt(target: Any) {
   def toBooleanOrNull(): JBoolean = ToBooleanConversionExt.convertOrNull(target)
 }
 
-object ToBooleanConversionExt extends ExtensionMethodsHandler with Conversion {
+object ToBooleanConversionExt extends ConversionExt {
   private val cannotConvertException = (value: Any) =>
     new IllegalArgumentException(s"Cannot convert: $value to Boolean")
   private val allowedClassesForConversion: Set[Class[_]] = Set(classOf[String], classOf[Object])
+  private val booleanTyping: typing.TypedClass           = Typed.typedClass[JBoolean]
 
-  private val definition = StaticMethodDefinition(
-    signature = MethodTypeInfo.noArgTypeInfo(Typed.typedClass[JBoolean]),
-    name = "",
-    description = None
+  override val definitions: List[StaticMethodDefinition] = List(
+    definition(booleanTyping, "isBoolean", Some("Check whether can be convert to a Boolean")),
+    definition(booleanTyping, "toBoolean", Some("Convert to Boolean or throw exception in case of failure")),
+    definition(booleanTyping, "toBooleanOrNull", Some("Convert to Boolean or null in case of failure")),
   )
-
-  private val definitions = List(
-    definition.copy(name = "isBoolean", description = Some("Check whether can be convert to a Boolean")),
-    definition.copy(name = "toBoolean", description = Some("Convert to Boolean or throw exception in case of failure")),
-    definition.copy(name = "toBooleanOrNull", description = Some("Convert to Boolean or null in case of failure")),
-  ).groupBy(_.name)
 
   override type ExtensionMethodInvocationTarget = ToBooleanConversionExt
   override val invocationTargetClass: Class[ToBooleanConversionExt] = classOf[ToBooleanConversionExt]
@@ -36,12 +31,6 @@ object ToBooleanConversionExt extends ExtensionMethodsHandler with Conversion {
       set: ClassDefinitionSet
   ): ToExtensionMethodInvocationTargetConverter[ToBooleanConversionExt] =
     (target: Any) => new ToBooleanConversionExt(target)
-
-  override def extractDefinitions(clazz: Class[_], set: ClassDefinitionSet): Map[String, List[MethodDefinition]] =
-    if (appliesToConversion(clazz)) definitions
-    else Map.empty
-
-  override def appliesToClassInRuntime(clazz: Class[_]): Boolean = true
 
   override type ResultType = JBoolean
   override val resultTypeClass: Class[JBoolean] = classOf[JBoolean]
