+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/extension/ToDoubleConversionExt.scala
@@ -14,17 +14,17 @@ class ToDoubleConversionExt(target: Any) {
   def toDoubleOrNull(): JDouble = ToDoubleConversionExt.convertOrNull(target)
 }
 
-object ToDoubleConversionExt extends ExtensionMethodsHandler with ToNumericConversionExt {
+object ToDoubleConversionExt extends ConversionExt with ToNumericConversion {
   override type ExtensionMethodInvocationTarget = ToDoubleConversionExt
   override val invocationTargetClass: Class[ToDoubleConversionExt] = classOf[ToDoubleConversionExt]
   override type ResultType = JDouble
   override val resultTypeClass: Class[JDouble] = classOf[JDouble]
 
-  override val definitions: Map[String, List[MethodDefinition]] = List(
+  override val definitions: List[MethodDefinition] = List(
     definition(Typed.typedClass[JBoolean], "isDouble", Some("Check whether can be convert to a Double")),
     definition(Typed.typedClass[JDouble], "toDouble", Some("Convert to Double or throw exception in case of failure")),
     definition(Typed.typedClass[JDouble], "toDoubleOrNull", Some("Convert to Double or null in case of failure")),
-  ).groupBy(_.name)
+  )
 
   override def createConverter(
       set: ClassDefinitionSet
